+++ services/rover/src/oci_cli_rover/rover_utils.py
@@ -1,6 +1,8 @@
 # coding: utf-8
 # Copyright (c) 2016, 2021, Oracle and/or its affiliates.  All rights reserved.
 # This software is dual-licensed to you under the Universal Permissive License (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl or Apache License 2.0 as shown at http://www.apache.org/licenses/LICENSE-2.0. You may choose either license.
+import random
+import string
 
 import click
 import six
@@ -92,3 +94,107 @@ def prompt_for_workload_delete():
 def modify_image_workload_name(image_workload_name):
     image_workload_name = image_workload_name.replace(" ", "")
     return image_workload_name
+
+
+def create_master_key_policy_rover_resource(resource_name, ctx, **kwargs):
+    confirm_prompt_policy = "You are providing your own master key ID, please create a policy to allow " \
+                            "Roving Edge Infrastructure to create DEKs and use the master key ID to encrypt " \
+                            "and decrypt data. Do you want to create the policy now?(Y/N)"
+    value_policy = click.confirm(click.style(confirm_prompt_policy, fg="yellow"))
+    if value_policy:
+        policy_kwargs = {'master_key_id': kwargs['master_key_id']}
+        #  Validate policy parameters. If policy-compartment-id is not provided,
+        #  Check if policy should be created in the same compartment as the rover resource
+        validate_policy_parameters(**kwargs)
+        if not kwargs['policy_compartment_id']:
+            confirm_policy_compartment = "Do you want to create the policy in the rover " + resource_name + \
+                                         "'s compartment?(Y/N)"
+            value_policy_compartment = click.confirm(click.style(confirm_policy_compartment, fg="yellow"))
+            if value_policy_compartment:
+                policy_kwargs['policy_compartment_id'] = kwargs['compartment_id']
+
+        # Pass policy name as display name if provided else generate one
+        if kwargs['policy_name']:
+            policy_kwargs['policy_name'] = kwargs['policy_name'].strip()
+        else:
+            policy_kwargs['policy_name'] = kwargs['display_name'].strip() + "-" + generate_random_string()
+        setup_master_key_policy(ctx, **policy_kwargs)
+    else:
+        click.echo("WARNING: Creation of the rover " + resource_name + " might fail if the policies are not provided")
+
+
+def setup_master_key_policy(ctx, **kwargs):
+    # Validate master-key-id
+    if isinstance(kwargs['master_key_id'], six.string_types) and len(kwargs['master_key_id'].strip()) == 0:
+        raise click.UsageError('Parameter master_key_id cannot be whitespace or empty string')
+    config = oci.config.from_file(file_location=ctx.obj['config_file'], profile_name=ctx.obj['profile'])
+    root_compartment = config['tenancy']
+    # Use compartment id if provided, otherwise Use root tenancy
+    if not kwargs['policy_compartment_id']:
+        click.echo("WARNING: policy-compartment-id not provided, policy will be created in root tenancy")
+        kwargs['policy_compartment_id'] = root_compartment
+
+    # Validate policy name is provided, else generate one
+    if not kwargs['policy_name']:
+        kwargs['policy_name'] = config['user'][-8:] + "-" + generate_random_string()
+
+    ctx.endpoint = None
+    ctx.obj['endpoint'] = None
+    compartment_id = kwargs['policy_compartment_id']
+    policy_name = kwargs['policy_name'] + "-master-key-id-policy"
+
+    # setting up home region and configuration in ctx since IAM resources can be created only in the home region of
+    # the tenancy
+    identity_client = cli_util.build_client('identity', 'identity', ctx)
+    subscription_kwargs = {}
+    subscriptions_result = identity_client.list_region_subscriptions(
+        tenancy_id=root_compartment, **subscription_kwargs)
+
+    if len(subscriptions_result.data) > 1:
+        for subscription in subscriptions_result.data:
+            if subscription.is_home_region:
+                ctx.obj['region'] = subscription.region_name
+                ctx.obj['config']['region'] = subscription.region_name
+            break
+
+    # Get identity client
+    identity_client = cli_util.build_client('identity', 'identity', ctx)
+
+    # setting up policies
+    click.echo('Setting up the policies in the compartment.')
+    statement_1 = "Allow service rover to use keys where target.key.id is " + kwargs['master_key_id']
+    iam_kwargs = {}
+    iam_details = {
+        'compartmentId': compartment_id,
+        'name': policy_name,
+        'statements': [statement_1],
+        'description': "The policies to use customer's master key"
+    }
+
+    create_policies_result = identity_client.create_policy(
+        create_policy_details=iam_details,
+        **iam_kwargs
+    )
+    return create_policies_result
+
+
+def validate_policy_parameters(**kwargs):
+    if 'policy_compartment_id' in kwargs and kwargs['policy_compartment_id']:
+        if isinstance(kwargs['policy_compartment_id'], six.string_types) and \
+                len(kwargs['policy_compartment_id'].strip()) == 0:
+            raise click.UsageError('Parameter policy-compartment-id cannot be whitespace or empty string')
+    if 'policy_name' in kwargs and kwargs['policy_name']:
+        if isinstance(kwargs['policy_name'], six.string_types) and len(kwargs['policy_name'].strip()) == 0:
+            raise click.UsageError('Policy name cannot be whitespace or empty string')
+
+
+def generate_random_string():
+    return "".join(random.choice(string.ascii_lowercase) for _ in range(10))
+
+
+def remove_additional_params_after_policy(**kwargs):
+    if 'policy_name' in kwargs:
+        kwargs.pop('policy_name')
+    if 'policy_compartment_id' in kwargs:
+        kwargs.pop('policy_compartment_id')
+    return kwargs